requirements.txt
Flask
biopython
reportlab
gunicorn
numpy
pandas
